update: rename Config to ChatTabConfig
