+++ simplecloud-modules/simplecloud-module-chat-tab/src/main/kotlin/eu/thesimplecloud/module/prefix/service/spigot/BukkitPluginMain.kt
@@ -23,9 +23,7 @@
 package eu.thesimplecloud.module.prefix.service.spigot
 
 import eu.thesimplecloud.api.CloudAPI
-import eu.thesimplecloud.module.prefix.config.Config
-import eu.thesimplecloud.module.prefix.service.command.ChatTabCommand
-import eu.thesimplecloud.module.prefix.service.spigot.configuration.DelayConfiguration
+import eu.thesimplecloud.module.prefix.manager.config.ChatTabConfig
 import eu.thesimplecloud.module.prefix.service.spigot.listener.ChatListener
 import eu.thesimplecloud.module.prefix.service.spigot.listener.CloudListener
 import eu.thesimplecloud.module.prefix.service.spigot.listener.JoinListener
@@ -42,20 +40,16 @@ import org.bukkit.plugin.java.JavaPlugin
  */
 class BukkitPluginMain : JavaPlugin() {
 
-    val delayConfiguration: DelayConfiguration = DelayConfiguration()
-
     override fun onEnable() {
 
         val groupName = CloudPlugin.instance.thisService().getGroupName()
-        if (Config.getConfig().disabledServerGroups.contains(groupName)) {
+        if (ChatTabConfig.getConfig().disabledServerGroups.contains(groupName)) {
             Bukkit.getLogger().info("[SimpleCloud] The Chat+Tab module is deactivated on this service!")
             Bukkit.getPluginManager().disablePlugin(this)
             return
         }
 
         TablistHelper.load()
-
-        getCommand("chat-tab")?.setExecutor(ChatTabCommand(this))
 
         Bukkit.getPluginManager().registerEvents(JoinListener(this), this)
         Bukkit.getPluginManager().registerEvents(ChatListener(), this)
