+++ simplecloud-modules/simplecloud-module-chat-tab/src/main/kotlin/eu/thesimplecloud/module/prefix/service/tablist/ProxyTablistHelper.kt
@@ -23,8 +23,8 @@
 package eu.thesimplecloud.module.prefix.service.tablist
 
 import eu.thesimplecloud.module.permission.PermissionPool
-import eu.thesimplecloud.module.prefix.config.Config
-import eu.thesimplecloud.module.prefix.config.TablistInformation
+import eu.thesimplecloud.module.prefix.manager.config.ChatTabConfig
+import eu.thesimplecloud.module.prefix.manager.config.TablistInformation
 import java.util.*
 
 /**
@@ -36,7 +36,7 @@ object ProxyTablistHelper {
         val permissionPlayer =
             PermissionPool.instance.getPermissionPlayerManager().getCachedPermissionPlayer(uuid) ?: return null
 
-        val informationList = Config.getConfig().informationList
+        val informationList = ChatTabConfig.getConfig().informationList
         val tablistInformation = informationList.sortedBy { it.priority }.first {
             permissionPlayer.hasPermissionGroup(it.groupName)
         }
